+++ fire_atlas_editor/lib/screens/editor_screen/widgets/selection_canvas/selection_canvas.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/material.dart' hide Image;
 import 'package:flame/flame.dart';
 import 'package:flame/sprite.dart';
 import '../../../../widgets/container.dart';
@@ -8,35 +8,48 @@ import '../../../../store/store.dart';
 
 import './canvas_board.dart';
 
+import 'dart:ui';
+
 class SelectionCanvas extends StatelessWidget {
   @override
   Widget build(_) {
     return MicroStoreProvider<FireAtlasState>(
-        store: Store.instance,
-        builder: (ctx, store) => FContainer(
-            child: FutureBuilder(
-                future: Flame.images.fromBase64(
-                  store.state.currentAtlas.id,
-                  store.state.currentAtlas.imageData,
-                ),
-                builder: (ctx, snapshot) {
-                  if (snapshot.hasData) {
-                    return LayoutBuilder(
-                      builder: (ctx, constraints) {
-                        final size =
-                            Size(constraints.maxWidth, constraints.maxHeight);
-                        return CanvasBoard(
-                          sprite: Sprite(snapshot.data),
-                          size: size,
-                          tileHeight: store.state.currentAtlas.tileHeight,
-                          tileWidth: store.state.currentAtlas.tileWidth,
-                        );
-                      },
+      store: Store.instance,
+      builder: (ctx, store) {
+        final atlas = store.state.currentAtlas;
+
+        if (atlas == null) {
+          return Text('No atlas selected');
+        }
+
+        return FContainer(
+          child: FutureBuilder<Image>(
+            future: Flame.images.fromBase64(
+              atlas.id,
+              atlas.imageData!,
+            ),
+            builder: (ctx, snapshot) {
+              if (snapshot.hasData) {
+                return LayoutBuilder(
+                  builder: (ctx, constraints) {
+                    final size = Size(
+                      constraints.maxWidth,
+                      constraints.maxHeight,
                     );
-                  } else if (snapshot.hasError) {
-                    return Text('Something wrong happened :(');
-                  }
-                  return Text('Something wrong happened :(');
-                })));
+                    return CanvasBoard(
+                      sprite: Sprite(snapshot.data!),
+                      size: size,
+                      tileHeight: atlas.tileHeight,
+                      tileWidth: atlas.tileWidth,
+                    );
+                  },
+                );
+              }
+              return Text('Something wrong happened :(');
+            },
+          ),
+        );
+      },
+    );
   }
 }