+++ flame_fire_atlas/example/lib/main.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:flame/anchor.dart';
-import 'package:flame/extensions/vector2.dart';
+import 'package:flame/extensions.dart';
 import 'package:flame/game.dart';
 import 'package:flame/gestures.dart';
-import 'package:flame/components/sprite_animation_component.dart';
-import 'package:flame/components/sprite_component.dart';
+import 'package:flame/components.dart';
 import 'package:flame_fire_atlas/flame_fire_atlas.dart';
 
 void main() async {
@@ -19,38 +17,38 @@ void main() async {
 }
 
 class ExampleGame extends BaseGame with TapDetector {
-  FireAtlas _atlas;
+  late FireAtlas _atlas;
 
   @override
   Future<void> onLoad() async {
     _atlas = await loadFireAtlas('caveace.fa');
     add(
       SpriteAnimationComponent(
-        Vector2(150, 100),
-        _atlas.getAnimation('shooting_ptero'),
+        size: Vector2(150, 100),
+        animation: _atlas.getAnimation('shooting_ptero'),
       )..y = 50,
     );
 
     add(SpriteAnimationComponent(
-      Vector2(150, 100),
-      _atlas.getAnimation('bomb_ptero'),
+      size: Vector2(150, 100),
+      animation: _atlas.getAnimation('bomb_ptero'),
     )
       ..y = 50
       ..x = 200);
 
     add(
-      SpriteComponent.fromSprite(Vector2(50, 50), _atlas.getSprite('bullet'))
+      SpriteComponent(size: Vector2(50, 50), sprite: _atlas.getSprite('bullet'))
         ..y = 200,
     );
 
     add(
-      SpriteComponent.fromSprite(Vector2(50, 50), _atlas.getSprite('shield'))
+      SpriteComponent(size: Vector2(50, 50), sprite: _atlas.getSprite('shield'))
         ..x = 100
         ..y = 200,
     );
 
     add(
-      SpriteComponent.fromSprite(Vector2(50, 50), _atlas.getSprite('ham'))
+      SpriteComponent(size: Vector2(50, 50), sprite: _atlas.getSprite('ham'))
         ..x = 200
         ..y = 200,
     );
@@ -61,8 +59,8 @@ class ExampleGame extends BaseGame with TapDetector {
     final o = details.localPosition;
 
     add(SpriteAnimationComponent(
-      Vector2(100, 100),
-      _atlas.getAnimation('explosion'),
+      size: Vector2(100, 100),
+      animation: _atlas.getAnimation('explosion'),
       removeOnFinish: true,
     )
       ..anchor = Anchor.center