+++ flame_fire_atlas/lib/flame_fire_atlas.dart
@@ -3,10 +3,8 @@ library flame_fire_atlas;
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/sprite.dart';
-import 'package:flame/extensions/vector2.dart';
-import 'package:flame/sprite_animation.dart';
-import 'package:flame/assets/images.dart';
-import 'package:flame/assets/assets_cache.dart';
+import 'package:flame/extensions.dart';
+import 'package:flame/assets.dart';
 
 import 'package:archive/archive.dart';
 
@@ -23,15 +21,45 @@ extension FireAtlasExtensions on Game {
   }
 }
 
-abstract class Selection {
+class Selection {
   String id;
   int x;
   int y;
   int w;
   int h;
 
+  Selection({
+    required this.id,
+    required this.x,
+    required this.y,
+    required this.w,
+    required this.h,
+  });
+
+  factory Selection.fromJson(Map<String, dynamic> json) {
+    return Selection(
+      id: json['id'],
+      x: json['x'],
+      y: json['y'],
+      w: json['w'],
+      h: json['h'],
+    );
+  }
+}
+
+abstract class BaseSelection {
+  final Selection _info;
+
+  BaseSelection(this._info);
+
+  String get id => _info.id;
+  int get x => _info.x;
+  int get y => _info.y;
+  int get w => _info.w;
+  int get h => _info.h;
+
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> json = {}
+    final json = <String, dynamic>{}
       ..['id'] = id
       ..['x'] = x
       ..['y'] = y
@@ -40,35 +68,46 @@ abstract class Selection {
 
     return json;
   }
-
-  fromJson(Map<String, dynamic> json) {
-    id = json['id'];
-    x = json['x'];
-    y = json['y'];
-    w = json['w'];
-    h = json['h'];
-  }
 }
 
-class SpriteSelection extends Selection {
+class SpriteSelection extends BaseSelection {
+  SpriteSelection({
+    required Selection info,
+  }) : super(info);
+
+  factory SpriteSelection.fromJson(Map<String, dynamic> json) {
+    final info = Selection.fromJson(json);
+    return SpriteSelection(info: info);
+  }
+
   @override
   Map<String, dynamic> toJson() {
     return super.toJson()..['type'] = 'sprite';
   }
 }
 
-class AnimationSelection extends Selection {
+class AnimationSelection extends BaseSelection {
   int frameCount;
   double stepTime;
   bool loop;
 
-  @override
-  void fromJson(Map<String, dynamic> json) {
-    super.fromJson(json);
+  AnimationSelection({
+    required Selection info,
+    required this.frameCount,
+    required this.stepTime,
+    required this.loop,
+  }) : super(info);
 
-    frameCount = json['frameCount'];
-    stepTime = json['stepTime'];
-    loop = json['loop'];
+  @override
+  factory AnimationSelection.fromJson(Map<String, dynamic> json) {
+    final info = Selection.fromJson(json);
+
+    return AnimationSelection(
+      info: info,
+      frameCount: json['frameCount'],
+      stepTime: json['stepTime'],
+      loop: json['loop'],
+    );
   }
 
   @override
@@ -85,19 +124,29 @@ class FireAtlas {
   String id;
   double tileWidth;
   double tileHeight;
-  String imageData;
-  Image _image;
+  String? imageData;
+  Image? _image;
 
-  Map<String, Selection> selections = {};
+  FireAtlas({
+    required this.id,
+    required this.tileWidth,
+    required this.tileHeight,
+    required this.imageData,
+  });
+
+  Map<String, BaseSelection> selections = {};
 
   /// Loads the atlas image into memory so it can be used
   ///
   /// [clearImageData] Can be set to false to avoid clearing the stored information about the image on this object, this is true by default, its use is intended to enable serializing this object
   /// [images] The images cache to be used, falls back to [Flame.images] when omitted
   ///
-  Future<void> loadImage({bool clearImageData = true, Images images}) async {
+  Future<void> loadImage({bool clearImageData = true, Images? images}) async {
+    if (imageData == null) {
+      throw 'Attempting on calling load on an already loaded Image';
+    }
     final _images = images ?? Flame.images;
-    _image = await _images.fromBase64(id, imageData);
+    _image = await _images.fromBase64(id, imageData!);
 
     // Clear memory
     if (clearImageData) {
@@ -106,7 +155,7 @@ class FireAtlas {
   }
 
   Map<String, dynamic> toJson() {
-    final Map<String, dynamic> selectionsJson = {};
+    final selectionsJson = <String, dynamic>{};
     selections.entries.forEach((entry) {
       selectionsJson[entry.key] = entry.value.toJson();
     });
@@ -115,27 +164,25 @@ class FireAtlas {
       ..['id'] = id
       ..['imageData'] = imageData
       ..['selections'] = selectionsJson
-      ..['tileWidth'] = tileWidth?.toDouble()
-      ..['tileHeight'] = tileHeight?.toDouble();
+      ..['tileWidth'] = tileWidth.toDouble()
+      ..['tileHeight'] = tileHeight.toDouble();
 
     return json;
   }
 
   static FireAtlas _fromJson(Map<String, dynamic> json) {
-    final atlas = FireAtlas()
-      ..id = json['id']
-      ..imageData = json['imageData']
-      ..tileHeight =
-          json['tileHeight']?.toDouble() ?? json['tileSize']?.toDouble()
-      ..tileWidth =
-          json['tileWidth']?.toDouble() ?? json['tileSize']?.toDouble();
+    final atlas = FireAtlas(
+      id: json['id'],
+      imageData: json['imageData'],
+      tileHeight:
+          json['tileHeight']?.toDouble() ?? json['tileSize']?.toDouble(),
+      tileWidth: json['tileWidth']?.toDouble() ?? json['tileSize']?.toDouble(),
+    );
 
     json['selections'].entries.forEach((entry) {
-      Selection selection = entry.value['type'] == 'animation'
-          ? AnimationSelection()
-          : SpriteSelection();
-
-      selection.fromJson(entry.value);
+      BaseSelection selection = entry.value['type'] == 'animation'
+          ? AnimationSelection.fromJson(entry.value)
+          : SpriteSelection.fromJson(entry.value);
 
       atlas.selections[entry.key] = selection;
     });
@@ -144,7 +191,7 @@ class FireAtlas {
   }
 
   static Future<FireAtlas> loadAsset(String fileName,
-      {AssetsCache assets, Images images}) async {
+      {AssetsCache? assets, Images? images}) async {
     final _assets = assets ?? Flame.assets;
 
     final bytes = await _assets.readBinaryFile(fileName);
@@ -157,7 +204,11 @@ class FireAtlas {
     String raw = jsonEncode(toJson());
 
     List<int> stringBytes = utf8.encode(raw);
-    List<int> gzipBytes = GZipEncoder().encode(stringBytes);
+    List<int>? gzipBytes = GZipEncoder().encode(stringBytes);
+
+    if (gzipBytes == null) {
+      throw 'Generated an empty file';
+    }
     return gzipBytes;
   }
 
@@ -167,24 +218,29 @@ class FireAtlas {
     return _fromJson(jsonDecode(unzipedString));
   }
 
-  void _assertImageLoaded() {
-    assert(
-      _image != null,
-      'Atlas is not loaded yet, call "load" before using it',
-    );
+  Image _assertImageLoaded() {
+    if (_image == null) {
+      throw 'Atlas is not loaded yet, call "load" before using it';
+    }
+
+    return _image!;
   }
 
   Sprite getSprite(String selectionId) {
     final selection = selections[selectionId];
 
-    _assertImageLoaded();
-    assert(selection != null,
-        'There is no selection with the id "$selectionId" on this atlas');
-    assert(selection is SpriteSelection,
-        'Selection "$selectionId" is not a Sprite');
+    if (selection == null) {
+      throw 'There is no selection with the id "$selectionId" on this atlas';
+    }
+
+    if (!(selection is SpriteSelection)) {
+      throw 'Selection "$selectionId" is not a Sprite';
+    }
+
+    final image = _assertImageLoaded();
 
     return Sprite(
-      _image,
+      image,
       srcPosition: Vector2(
         selection.x.toDouble() * tileWidth,
         selection.y.toDouble() * tileHeight,
@@ -199,26 +255,26 @@ class FireAtlas {
   SpriteAnimation getAnimation(String selectionId) {
     final selection = selections[selectionId];
 
-    _assertImageLoaded();
-    assert(selection != null,
-        'There is no selection with the id "$selectionId" on this atlas');
-    assert(selection is AnimationSelection,
-        'Selection "$selectionId" is not an Animation');
+    final image = _assertImageLoaded();
+
+    if (selection == null) {
+      throw 'There is no selection with the id "$selectionId" on this atlas';
+    }
+    if (!(selection is AnimationSelection)) {
+      throw 'Selection "$selectionId" is not an Animation';
+    }
 
     final initialX = selection.x.toDouble();
 
-    final animationSelection = selection as AnimationSelection;
-
-    final frameSize =
-        (1 + selection.w.toDouble()) / animationSelection.frameCount;
+    final frameSize = (1 + selection.w.toDouble()) / selection.frameCount;
 
     final width = frameSize * tileWidth;
     final height = (1 + selection.h.toDouble()) * tileHeight;
 
-    final sprites = List.generate(animationSelection.frameCount, (i) {
+    final sprites = List.generate(selection.frameCount, (i) {
       final x = (initialX + i) * frameSize;
       return Sprite(
-        _image,
+        image,
         srcPosition: Vector2(
           x * tileWidth,
           selection.y.toDouble() * tileHeight,
@@ -229,8 +285,8 @@ class FireAtlas {
 
     return SpriteAnimation.spriteList(
       sprites,
-      stepTime: animationSelection.stepTime,
-      loop: animationSelection.loop,
+      stepTime: selection.stepTime,
+      loop: selection.loop,
     );
   }
 }