+++ flame_fire_atlas/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [next]
+
+ * Compatibility with Flame v1.0.0-rc9 and null safe support
+
 ## [1.0.0-rc1]
 
  * Compatibility with Flame v1.0.0-rc5